Accept of selector syntax (with no effect)
