+++ packages/alfa-selector/src/selector/simple/pseudo-class/nth-child.ts
@@ -1,6 +1,9 @@
-import type { Nth } from "@siteimprove/alfa-css";
+import { Function, Parser as CSSParser, Token } from "@siteimprove/alfa-css";
+import { Nth } from "@siteimprove/alfa-css";
 import { Element } from "@siteimprove/alfa-dom";
-import { Option } from "@siteimprove/alfa-option";
+import { Maybe, Option } from "@siteimprove/alfa-option";
+import { Parser } from "@siteimprove/alfa-parser";
+import { Thunk } from "@siteimprove/alfa-thunk";
 
 import { Context } from "../../../context";
 
@@ -9,6 +12,8 @@ import { type Absolute, Universal } from "../../index";
 import { WithIndex } from "./pseudo-class";
 
 const { isElement } = Element;
+const { delimited, end, left, map, option, pair, right } = Parser;
+const { parseColon, parseIdent, parseWhitespace } = Token;
 
 /**
  * {@link https://drafts.csswg.org/selectors/#nth-child-pseudo}
@@ -16,9 +21,12 @@ const { isElement } = Element;
 export class NthChild extends WithIndex<"nth-child"> {
   public static of(
     index: Nth,
-    selector: Absolute = Universal.of(Option.of("*")),
+    selector: Maybe<Absolute> = Universal.of(Option.of("*")),
   ): NthChild {
-    return new NthChild(index, selector);
+    return new NthChild(
+      index,
+      Maybe.toOption(selector).getOr(Universal.of(Option.of("*"))),
+    );
   }
 
   private _selector: Absolute;
@@ -68,5 +76,30 @@ export class NthChild extends WithIndex<"nth-child"> {
 export namespace NthChild {
   export interface JSON extends WithIndex.JSON<"nth-child"> {}
 
-  export const parse = WithIndex.parseWithIndex("nth-child", NthChild.of);
+  const parseNth = left(
+    Nth.parse,
+    end((token) => `Unexpected token ${token}`),
+  );
+
+  export function parse(
+    parseSelector: Thunk<CSSParser<Absolute>>,
+  ): CSSParser<NthChild> {
+    return map(
+      right(
+        parseColon,
+        Function.parse("nth-child", () =>
+          pair(
+            Nth.parse,
+            option(
+              right(
+                delimited(parseWhitespace, parseIdent("of")),
+                parseSelector(),
+              ),
+            ),
+          ),
+        ),
+      ),
+      ([, [nth, selector]]) => NthChild.of(nth, selector),
+    );
+  }
 }
