+++ packages/alfa-selector/test/pseudo-class.spec.tsx
@@ -32,50 +32,6 @@ test(".parse() parses a functional pseudo-class selector", (t) => {
   });
 });
 
-test("#matches() checks if an element matches an :nth-child selector", (t) => {
-  const selector = parse(":nth-child(odd)");
-
-  const a = <p />;
-  const b = <p />;
-  const c = <p />;
-  const d = <p />;
-
-  <div>
-    {a}
-    Hello
-    {b}
-    {c}
-    {d}
-  </div>;
-
-  t.equal(selector.matches(a), true);
-  t.equal(selector.matches(b), false);
-  t.equal(selector.matches(c), true);
-  t.equal(selector.matches(d), false);
-});
-
-test("#matches() checks if an element matches an :nth-last-child selector", (t) => {
-  const selector = parse(":nth-last-child(odd)");
-
-  const a = <p />;
-  const b = <p />;
-  const c = <p />;
-  const d = <p />;
-
-  <div>
-    {a}
-    Hello
-    {b}
-    {c}
-    {d}
-  </div>;
-
-  t.equal(selector.matches(a), false);
-  t.equal(selector.matches(b), true);
-  t.equal(selector.matches(c), false);
-  t.equal(selector.matches(d), true);
-});
-
 test("#matches() checks if an element matches a :first-child selector", (t) => {
   const selector = parse(":first-child");
 
@@ -329,15 +285,4 @@ test("#matches() checks if an element matches a :visited selector", (t) => {
   for (const element of [<a />, <p />]) {
     t.equal(selector.matches(element), false, element.toString());
   }
-});
-test(".parse() parses an :nth-child selector", (t) => {
-  t.deepEqual(serialize(":nth-child(odd)"), {
-    type: "pseudo-class",
-    name: "nth-child",
-    index: {
-      step: 2,
-      offset: 1,
-    },
-    specificity: { a: 0, b: 1, c: 0 },
-  });
 });
